game works and ends
