+++ src/userTest.js
@@ -9,7 +9,7 @@ const chalk = require('chalk');
 clear();
 console.log(
   chalk.blueBright(
-figlet.textSync('SUPERTYPE :   REVOLUTION', {font:'ANSI Shadow', horizontalLayout: 'full' })
+    figlet.textSync('SUPERTYPE :   REVOLUTION', {font:'ANSI Shadow', horizontalLayout: 'full' })
   )
 );
